[TM-2855] initial refactor
