+++ src/components/elements/Map-mapbox/Map.d.ts
@@ -14,3 +14,10 @@ export interface LayerType {
 export type ControlType = Control | IControl;
 
 export type TooltipType = "edit" | "goTo" | "view";
+
+export type DashboardGetProjectsData = {
+  uuid?: string;
+  name?: string;
+  lat?: number;
+  long?: number;
+};
