+++ src/components/elements/Map-mapbox/utils.ts
@@ -1,1197 +1,40 @@
-import MapboxDraw from "@mapbox/mapbox-gl-draw";
-import bbox from "@turf/bbox";
-import * as turfHelper from "@turf/helpers";
-import _ from "lodash";
-import mapboxgl, { LngLat } from "mapbox-gl";
-import { createElement } from "react";
-import { createRoot } from "react-dom/client";
+/**
+ * utils.ts — backward-compat re-export barrel.
+ *
+ * All map logic has been extracted into domain files:
+ *   adapters/geojson.ts      — GeoJSON conversion helpers
+ *   adapters/geoserver.ts    — Geoserver URL builders + ANR layer constants
+ *   adapters/camera.ts       — Camera movement (zoom / center / marker)
+ *   layers/polygonLayers.ts  — Polygon source/layer/filter management
+ *   layers/overlayLayers.ts  — Borders, ANR overlay, basemap helpers
+ *   layers/mediaLayers.ts    — Media source/layer management
+ *   interactions/draw.ts     — Draw mode + polygon save operations
+ *   interactions/popups.ts   — Popup management (instance-scoped)
+ *
+ * New code should import directly from those files.
+ * This barrel preserves backward compat for existing callers.
+ */
 
+// ── Domain re-exports ──────────────────────────────────────────────────────────
+export * from "./adapters/geojson";
+export * from "./adapters/geoserver";
+export * from "./adapters/camera";
+export * from "./layers/polygonLayers";
+export * from "./layers/overlayLayers";
+export * from "./layers/mediaLayers";
+export * from "./interactions/draw";
+export * from "./interactions/popups";
+
+// ── Kept here: download / format / validation helpers ─────────────────────────
 import {
   loadPolygonGeoJson,
   loadProjectPolygonsGeoJson,
   loadProjectSitePolygonsGeoJson,
   loadSitePolygonsGeoJson
 } from "@/connections/GeoJsonExport";
-import { createProjectPolygonWithReplace, updateProjectPolygonResource } from "@/connections/ProjectPolygons";
-import { createSitePolygonsResource } from "@/connections/SitePolygons";
-import { geoserverUrl, geoserverWorkspace } from "@/constants/environment";
-import { LAYERS_NAMES, layersList } from "@/constants/layers";
-import { MediaDto } from "@/generated/v3/entityService/entityServiceSchemas";
 import { GetSitePolygonsGeoJsonQueryParams } from "@/generated/v3/researchService/researchServiceComponents";
-import { GeoJsonExportDto } from "@/generated/v3/researchService/researchServiceSchemas";
-import {
-  CreateSitePolygonAttributesDto,
-  SitePolygonLightDto
-} from "@/generated/v3/researchService/researchServiceSchemas";
+import { GeoJsonExportDto, SitePolygonLightDto } from "@/generated/v3/researchService/researchServiceSchemas";
 import Log from "@/utils/log";
-
-import { AnrPlotMapPopup } from "./components/AnrPlotMapPopup";
-import { MediaPopup } from "./components/MediaPopup";
-import { BBox, Feature, FeatureCollection, GeoJsonProperties, Geometry } from "./GeoJSON";
-import { DashboardGetProjectsData } from "./Map";
-import type { LayerType, LayerWithStyle, TooltipType } from "./Map.d";
-import { BASEMAP_CONFIGS, MapStyle } from "./MapControls/types";
-import { getPulsingDot } from "./pulsing.dot";
-
-/** Prefix must stay in sync with Google satellite visibility helpers. */
-export const ANR_PLOT_SOURCE_ID = "anr_plot_geometry-source";
-export const ANR_PLOT_FILL_LAYER_ID = "anr_plot_geometry-fill";
-export const ANR_PLOT_LINE_LAYER_ID = "anr_plot_geometry-line";
-export const ANR_PLOT_LAYER_PREFIX = "anr_plot_geometry";
-
-type DataPolygonOverview = {
-  status: string;
-  count: number;
-}[];
-
-export const getFeatureProperties = <T extends any>(properties: any, key: string): T | undefined => {
-  return properties[key] ?? properties[`user_${key}`];
-};
-
-export const convertToGeoJSON = (featureCollection: FeatureCollection) => {
-  const { features } = featureCollection;
-  return features.reduce((acc: turfHelper.Feature[], feature) => {
-    const { geometry, properties } = feature;
-    const coordinates = geometry.coordinates;
-    const type = geometry.type;
-    const featureGeoJSON = turfHelper.feature({ type, coordinates }, properties);
-    acc.push(featureGeoJSON);
-    return acc;
-  }, []);
-};
-
-export const convertToAcceptedGEOJSON = (geojson: any) => {
-  if (geojson.type !== "FeatureCollection" && geojson.type !== "GeometryCollection") {
-    const templateGeoJSON = {
-      type: "Feature",
-      properties: {},
-      geometry: geojson
-    };
-    const geojsonFormatted = {
-      type: "FeatureCollection",
-      features: [templateGeoJSON]
-    };
-    return geojsonFormatted;
-  } else if (geojson.type === "GeometryCollection") {
-    const geojsonFormatted = {
-      type: "FeatureCollection",
-      features: geojson.geometries.map((geometry: any) => ({
-        type: "Feature",
-        properties: {},
-        geometry
-      }))
-    };
-    return geojsonFormatted;
-  }
-  return geojson;
-};
-
-export const startDrawing = (draw: MapboxDraw, map: mapboxgl.Map) => {
-  draw.changeMode("draw_polygon");
-  map.getCanvas().style.cursor = "crosshair";
-};
-
-export const stopDrawing = (draw: MapboxDraw, map: mapboxgl.Map) => {
-  draw.changeMode("simple_select");
-  map.getCanvas().style.cursor = "auto";
-};
-
-export const addFilterOnLayer = (layer: any, parsedPolygonData: Record<string, string[]>, map: mapboxgl.Map) => {
-  addSourceToLayer(layer, map, parsedPolygonData);
-};
-const showPolygons = (
-  styles: LayerWithStyle[],
-  name: string,
-  map: mapboxgl.Map,
-  field: string,
-  parsedPolygonData: any,
-  zoomFilter?: number | undefined
-) => {
-  styles.forEach((style: LayerWithStyle, index: number) => {
-    const layerName = `${name}-${index}`;
-    if (!map.getLayer(layerName)) {
-      Log.warn(`Layer ${layerName} does not exist.`);
-      return;
-    }
-    const polygonStatus = style?.metadata?.polygonStatus;
-    const uuidFilter = [
-      "in",
-      ["get", field],
-      ["literal", parsedPolygonData?.[polygonStatus] === undefined ? "" : parsedPolygonData[polygonStatus]]
-    ];
-
-    const completeFilter = zoomFilter ? ["all", uuidFilter, [">", ["zoom"], zoomFilter]] : ["all", uuidFilter];
-
-    map.setFilter(layerName, completeFilter);
-    map.setLayoutProperty(layerName, "visibility", "visible");
-  });
-};
-
-let popupAttachedMap: Record<string, mapboxgl.Popup[]> = {
-  POLYGON: [],
-  MEDIA: []
-};
-
-export const loadLayersInMap = (
-  map: mapboxgl.Map,
-  polygonsData: Record<string, string[]> | undefined,
-  layer: any,
-  zoomFilter?: number | undefined
-) => {
-  if (map) {
-    showPolygons(layer.styles, layer.name, map, "uuid", polygonsData, zoomFilter);
-  }
-};
-
-const handleLayerClick = (
-  e: any,
-  PopupComponent: any,
-  map: mapboxgl.Map,
-  setPolygonFromMap: any,
-  sitePolygonData: SitePolygonLightDto[] | undefined,
-  type: TooltipType,
-  editPolygon: { isOpen: boolean; uuid: string; primary_uuid?: string },
-  setEditPolygon: (value: { isOpen: boolean; uuid: string; primary_uuid?: string }) => void,
-  layerName?: string,
-  isDashboard?: string | undefined,
-  setFilters?: any,
-  dashboardCountries?: any,
-  setLoader?: (value: boolean) => void,
-  setMobilePopupData?: (value: any) => void
-) => {
-  const { lngLat, features } = e;
-  const feature = features?.[0];
-  if (!feature) {
-    Log.warn("No feature found in click event");
-    return;
-  }
-
-  if (setMobilePopupData && isDashboard) {
-    const popupData = {
-      feature,
-      layerName,
-      type,
-      setPolygonFromMap,
-      sitePolygonData,
-      isDashboard,
-      editPolygon,
-      setEditPolygon,
-      setFilters,
-      dashboardCountries
-    };
-    setMobilePopupData(popupData);
-    return;
-  }
-
-  if (layerName === LAYERS_NAMES.POLYGON_GEOMETRY && map.getLayer(ANR_PLOT_FILL_LAYER_ID) != null) {
-    const anrHits = map.queryRenderedFeatures(e.point, { layers: [ANR_PLOT_FILL_LAYER_ID] });
-    if (anrHits.length > 0) {
-      return;
-    }
-  }
-
-  removePopups("POLYGON");
-  const isCentroidLayer = layerName === LAYERS_NAMES.CENTROIDS;
-  const popupOptions: mapboxgl.PopupOptions = {
-    className: isCentroidLayer ? "popup-map no-tip" : "popup-map",
-    offset: isCentroidLayer
-      ? {
-          top: [0, 9],
-          "top-left": [0, 9],
-          "top-right": [0, 9],
-          bottom: [0, -9],
-          "bottom-left": [0, -9],
-          "bottom-right": [0, -9],
-          left: [9, 0],
-          right: [-9, 0]
-        }
-      : 0
-  };
-
-  const popupContent = document.createElement("div");
-  popupContent.className = "popup-content-map";
-  const root = createRoot(popupContent);
-
-  const createPopup = (lngLat: mapboxgl.LngLat) =>
-    new mapboxgl.Popup(popupOptions).setLngLat(lngLat).setDOMContent(popupContent);
-
-  const newPopup = createPopup(lngLat);
-
-  const commonProps = {
-    feature,
-    popup: newPopup,
-    setPolygonFromMap,
-    sitePolygonData,
-    type,
-    editPolygon,
-    setEditPolygon
-  };
-  newPopup.addTo(map);
-  popupAttachedMap["POLYGON"].push(newPopup);
-  root.render(createElement(PopupComponent, commonProps));
-};
-
-export const removePopups = (key: "POLYGON" | "MEDIA") => {
-  if (!popupAttachedMap[key]) return;
-  popupAttachedMap[key].forEach(popup => {
-    popup.remove();
-  });
-  popupAttachedMap[key] = [];
-};
-
-export const removeMediaLayer = (map: mapboxgl.Map) => {
-  const layerName = LAYERS_NAMES.MEDIA_IMAGES;
-  map.getLayer(layerName) && map.removeLayer(layerName);
-  map.getSource(layerName) && map.removeSource(layerName);
-  map.hasImage("pulsing-dot") && map.removeImage("pulsing-dot");
-};
-
-export const addFilterOfPolygonsData = (map: mapboxgl.Map, polygonsData: Record<string, string[]> | undefined) => {
-  if (map && polygonsData) {
-    if (map.isStyleLoaded() || map.loaded()) {
-      layersList.forEach((layer: LayerType) => {
-        loadLayersInMap(map, polygonsData, layer);
-      });
-    } else {
-      map.on("style.load", () => {
-        layersList.forEach((layer: LayerType) => {
-          loadLayersInMap(map, polygonsData, layer);
-        });
-      });
-      map.on("load", () => {
-        layersList.forEach((layer: LayerType) => {
-          loadLayersInMap(map, polygonsData, layer);
-        });
-      });
-    }
-  }
-};
-
-export const addGeojsonToDraw = (
-  geojson: any,
-  uuid: string,
-  cb: Function,
-  currentDraw: MapboxDraw,
-  map?: mapboxgl.Map
-) => {
-  if (geojson) {
-    const geojsonFormatted = convertToAcceptedGEOJSON(geojson);
-    const addToDrawAndFilter = () => {
-      if (currentDraw) {
-        currentDraw.add(geojsonFormatted);
-        const currentDrawFeatures = currentDraw.getAll();
-        currentDraw.set(currentDrawFeatures);
-        const featureId = currentDrawFeatures.features[0].id;
-        currentDraw.changeMode("direct_select", { featureId: featureId as string });
-        if (map) {
-          zoomToBbox(bbox(geojsonFormatted) as BBox, map, false);
-        }
-        cb(uuid);
-      }
-    };
-    addToDrawAndFilter();
-  }
-};
-
-export const addMediaSourceAndLayer = (
-  map: mapboxgl.Map,
-  modelFilesData: MediaDto[],
-  setImageCover: Function,
-  handleDownload: Function,
-  handleDelete: Function,
-  openModalImageDetail: Function,
-  isProjectPath: boolean
-) => {
-  const layerName = LAYERS_NAMES.MEDIA_IMAGES;
-  removeMediaLayer(map);
-  removePopups("MEDIA");
-  const modelFilesGeolocalized = modelFilesData!.filter(modelFile => modelFile.lat != null && modelFile.lng != null);
-  if (modelFilesGeolocalized.length === 0) {
-    return;
-  }
-
-  const features: Feature<Geometry, GeoJsonProperties>[] = modelFilesGeolocalized.map(modelFile => ({
-    type: "Feature",
-    geometry: {
-      type: "Point",
-      coordinates: [modelFile.lng, modelFile.lat]
-    },
-    properties: {
-      uuid: modelFile.uuid,
-      name: modelFile.name,
-      created_date: modelFile.createdAt,
-      thumbUrl: modelFile.thumbUrl,
-      location: {
-        lat: modelFile.lat,
-        lng: modelFile.lng
-      },
-      is_cover: modelFile.isCover,
-      is_public: modelFile.isPublic,
-      photographer: modelFile.photographer,
-      description: modelFile.description,
-      mime_type: modelFile.mimeType,
-      file_name: modelFile.fileName
-    }
-  }));
-
-  const pulsingDot = getPulsingDot(map, 120);
-
-  if (!map.hasImage("pulsing-dot")) {
-    map.addImage("pulsing-dot", pulsingDot, { pixelRatio: 4 });
-  }
-
-  map.addSource(layerName, {
-    type: "geojson",
-    data: {
-      type: "FeatureCollection",
-      features
-    }
-  });
-
-  map.addLayer({
-    id: layerName,
-    type: "symbol",
-    source: layerName,
-    layout: {
-      "icon-image": "pulsing-dot"
-    }
-  });
-
-  map.moveLayer(layerName);
-
-  map.on("click", layerName, e => {
-    e.preventDefault();
-    e.features!.forEach((feature: any) => {
-      let popupContent = document.createElement("div");
-      popupContent.className = "popup-content-media";
-      const root = createRoot(popupContent);
-
-      const coverImage = () => {
-        setImageCover(feature?.properties?.uuid);
-      };
-      const handleDownloadFunction = () => {
-        handleDownload(feature?.properties?.uuid, feature?.properties?.name);
-      };
-      const handleDeleteFunction = () => {
-        handleDelete(feature?.properties?.uuid);
-      };
-      const openModalImageDetailFunction = () => {
-        openModalImageDetail(feature?.properties);
-      };
-      root.render(
-        createElement(MediaPopup, {
-          ...feature.properties,
-          onClose: () => {
-            removePopups("MEDIA");
-          },
-          handleDownload: handleDownloadFunction,
-          coverImage: coverImage,
-          handleDelete: handleDeleteFunction,
-          openModalImageDetail: openModalImageDetailFunction,
-          isProjectPath: isProjectPath
-        })
-      );
-      const mediaPopup = new mapboxgl.Popup({ className: "popup-media", closeButton: false })
-        .setLngLat(feature.geometry.coordinates)
-        .setDOMContent(popupContent)
-        .addTo(map);
-      popupAttachedMap["MEDIA"].push(mediaPopup);
-    });
-  });
-};
-
-export const addSourcesToLayers = (
-  map: mapboxgl.Map,
-  polygonsData: Record<string, string[]> | undefined,
-  centroids: DashboardGetProjectsData[] | undefined,
-  zoomFilter?: number | undefined,
-  isDashboard?: string | undefined,
-  polygonsCentroids?: any[] | undefined
-) => {
-  if (map) {
-    layersList.forEach((layer: LayerType) => {
-      if (layer.name === LAYERS_NAMES.POLYGON_GEOMETRY) {
-        addSourceToLayer(layer, map, polygonsData, zoomFilter, isDashboard);
-      }
-      if (layer.name === LAYERS_NAMES.WORLD_COUNTRIES && isDashboard) {
-        addSourceToLayer(layer, map, undefined, undefined);
-      }
-      if (layer.name === LAYERS_NAMES.CENTROIDS && isDashboard) {
-        addGeojsonSourceToLayer(centroids, map, layer, zoomFilter, !_.isEmpty(polygonsData));
-      }
-    });
-    if (isDashboard) {
-      addPolygonCentroidsLayer(map, polygonsCentroids ?? [], zoomFilter);
-    }
-  }
-};
-export const addPolygonCentroidsLayer = (
-  map: mapboxgl.Map,
-  centroids: { uuid: string; long: number; lat: number }[],
-  zoomFilterValue?: number
-) => {
-  const layerName = LAYERS_NAMES.POLYGON_CENTROIDS;
-
-  if (!map) {
-    return;
-  }
-
-  if (!centroids || centroids.length === 0) {
-    return;
-  }
-
-  const validCentroids = centroids.filter(c => {
-    return (
-      c != null &&
-      typeof c.lat === "number" &&
-      !isNaN(c.lat) &&
-      typeof c.long === "number" &&
-      !isNaN(c.long) &&
-      typeof c.uuid === "string" &&
-      c.uuid.length > 0
-    );
-  });
-
-  if (validCentroids.length === 0) {
-    return;
-  }
-
-  let rafId: number | null = null;
-  let isActive = true;
-  let styleLoadHandler: (() => void) | null = null;
-
-  const addLayerToMap = (): boolean => {
-    if (!isActive || !map) return false;
-
-    if (!map.isStyleLoaded()) {
-      return false;
-    }
-
-    try {
-      if (map.getLayer(`${layerName}`)) {
-        map.removeLayer(`${layerName}`);
-      }
-      if (map.getSource(layerName)) {
-        map.removeSource(layerName);
-      }
-
-      if (map.hasImage("pulsing-dot-centroids")) {
-        map.removeImage("pulsing-dot-centroids");
-      }
-
-      const features: GeoJSON.Feature[] = validCentroids.map(centroid => ({
-        type: "Feature",
-        geometry: {
-          type: "Point",
-          coordinates: [centroid.long, centroid.lat]
-        },
-        properties: {
-          uuid: centroid.uuid
-        }
-      }));
-
-      const pulsingDot = getPulsingDot(map, 120, "#72D961");
-      map.addImage("pulsing-dot-centroids", pulsingDot, { pixelRatio: 4 });
-
-      map.addSource(layerName, {
-        type: "geojson",
-        data: {
-          type: "FeatureCollection",
-          features: features
-        }
-      });
-
-      const filter = zoomFilterValue ? ["<=", ["zoom"], zoomFilterValue] : [">=", ["zoom"], 0];
-      map.addLayer({
-        id: layerName,
-        type: "symbol",
-        source: layerName,
-        layout: {
-          "icon-image": "pulsing-dot-centroids"
-        },
-        paint: {},
-        filter: filter
-      });
-
-      return true;
-    } catch (error) {
-      Log.error("Error adding polygon centroids layer:", error);
-      return false;
-    }
-  };
-
-  const pollForStyleLoad = (attemptsLeft = 180) => {
-    if (!isActive) return;
-    if (!map) return;
-
-    const layerAdded = addLayerToMap();
-
-    if (layerAdded) {
-      if (rafId != null) {
-        cancelAnimationFrame(rafId);
-        rafId = null;
-      }
-      if (styleLoadHandler && map) {
-        map.off("style.load", styleLoadHandler);
-        styleLoadHandler = null;
-      }
-      return;
-    }
-
-    if (attemptsLeft > 0) {
-      rafId = requestAnimationFrame(() => pollForStyleLoad(attemptsLeft - 1));
-    } else {
-      Log.error("Failed to add polygon centroids layer after 180 requestAnimationFrame attempts");
-      if (styleLoadHandler && map) {
-        map.off("style.load", styleLoadHandler);
-        styleLoadHandler = null;
-      }
-    }
-  };
-
-  const layerAdded = addLayerToMap();
-  if (layerAdded) {
-    return;
-  }
-
-  pollForStyleLoad();
-
-  styleLoadHandler = () => {
-    if (isActive && map) {
-      const added = addLayerToMap();
-      if (added) {
-        if (rafId != null) {
-          cancelAnimationFrame(rafId);
-          rafId = null;
-        }
-        if (styleLoadHandler && map) {
-          map.off("style.load", styleLoadHandler);
-          styleLoadHandler = null;
-        }
-      }
-    }
-  };
-
-  map.on("style.load", styleLoadHandler);
-};
-export const addPopupsToMap = (
-  map: mapboxgl.Map,
-  popupComponent: any,
-  setPolygonFromMap: any,
-  sitePolygonData: SitePolygonLightDto[] | undefined,
-  type: TooltipType,
-  editPolygon: { isOpen: boolean; uuid: string; primary_uuid?: string },
-  setEditPolygon: (value: { isOpen: boolean; uuid: string; primary_uuid?: string }) => void,
-  draw: MapboxDraw,
-  isDashboard?: string | undefined,
-  setFilters?: any,
-  dashboardCountries?: any,
-  setLoader?: (value: boolean) => void,
-  selectedCountry?: string | null,
-  setMobilePopupData?: (value: any) => void
-) => {
-  if (popupComponent) {
-    layersList.forEach((layer: LayerType) => {
-      addPopupToLayer(
-        map,
-        popupComponent,
-        layer,
-        setPolygonFromMap,
-        sitePolygonData,
-        type,
-        editPolygon,
-        setEditPolygon,
-        draw,
-        isDashboard,
-        setFilters,
-        dashboardCountries,
-        setLoader,
-        selectedCountry,
-        setMobilePopupData
-      );
-    });
-  }
-};
-const activeClickHandlers: Record<string, any> = {};
-
-export const addPopupToLayer = (
-  map: mapboxgl.Map,
-  popupComponent: any,
-  layer: any,
-  setPolygonFromMap: any,
-  sitePolygonData: SitePolygonLightDto[] | undefined,
-  type: TooltipType,
-  editPolygon: { isOpen: boolean; uuid: string; primary_uuid?: string },
-  setEditPolygon: (value: { isOpen: boolean; uuid: string; primary_uuid?: string }) => void,
-  draw: MapboxDraw,
-  isDashboard?: string | undefined,
-  setFilters?: any,
-  dashboardCountries?: any,
-  setLoader?: (value: boolean) => void,
-  selectedCountry?: string | null,
-  setMobilePopupData?: (value: any) => void
-) => {
-  if (!popupComponent) return;
-
-  const { name } = layer;
-  let layers = map.getStyle().layers;
-  let targetLayers = layers.filter(layer => layer.id.startsWith(name));
-
-  if (name === LAYERS_NAMES.CENTROIDS && targetLayers.length > 0) {
-    targetLayers = targetLayers.filter(layer => (layer as any)?.metadata?.type === "big-circle");
-  }
-
-  const clickHandler = (e: any) => {
-    const currentMode = draw?.getMode();
-    if (currentMode === "draw_polygon" || currentMode === "draw_line_string") return;
-    if (name === LAYERS_NAMES.WORLD_COUNTRIES) return;
-
-    handleLayerClick(
-      e,
-      popupComponent,
-      map,
-      setPolygonFromMap,
-      sitePolygonData,
-      type,
-      editPolygon,
-      setEditPolygon,
-      name,
-      isDashboard,
-      setFilters,
-      dashboardCountries,
-      setLoader,
-      setMobilePopupData
-    );
-  };
-
-  targetLayers.forEach(targetLayer => {
-    if (activeClickHandlers[targetLayer.id]) {
-      map.off("click", targetLayer.id, activeClickHandlers[targetLayer.id]);
-      map.off("touchend", targetLayer.id, activeClickHandlers[targetLayer.id]);
-      delete activeClickHandlers[targetLayer.id];
-    }
-
-    activeClickHandlers[targetLayer.id] = clickHandler;
-    map.on("click", targetLayer.id, clickHandler);
-    map.on("touchend", targetLayer.id, clickHandler);
-  });
-};
-
-const getGeoserverURL = (layerName: string, isDashboard?: string | undefined) => {
-  const workspace = isDashboard ? `${geoserverWorkspace}_db` : geoserverWorkspace;
-  return `${geoserverUrl}/geoserver/gwc/service/wmts?REQUEST=GetTile&SERVICE=WMTS
-      &VERSION=1.0.0&LAYER=${workspace}:${layerName}&STYLE=&TILEMATRIX=EPSG:900913:{z}&TILEMATRIXSET=EPSG:900913&FORMAT=application/vnd.mapbox-vector-tile&TILECOL={x}&TILEROW={y}&RND=${Math.random()}`;
-};
-export const addHoverEvent = (layer: LayerType, map: mapboxgl.Map) => {
-  const { name, styles } = layer;
-  const layersToHover = styles.map((_, index) => `${name}-${index}`);
-  if (name === LAYERS_NAMES.WORLD_COUNTRIES) {
-    let hoveredPolygonId: any = null;
-    map.on("mousemove", layersToHover, e => {
-      const zoomLevel = map.getZoom();
-      if (zoomLevel <= 4.5) {
-        if (e.features && e.features.length > 0) {
-          if (hoveredPolygonId !== null) {
-            map.setFeatureState({ source: name, sourceLayer: name, id: hoveredPolygonId }, { hover: false });
-          }
-          hoveredPolygonId = e.features[0].id;
-          map.setFeatureState({ source: name, sourceLayer: name, id: hoveredPolygonId }, { hover: true });
-        }
-      } else {
-        if (hoveredPolygonId !== null) {
-          map.setFeatureState({ source: name, sourceLayer: name, id: hoveredPolygonId }, { hover: false });
-        }
-        hoveredPolygonId = null;
-      }
-    });
-    map.on("mouseleave", layersToHover, () => {
-      if (hoveredPolygonId !== null) {
-        map.setFeatureState({ source: name, sourceLayer: name, id: hoveredPolygonId }, { hover: false });
-      }
-      hoveredPolygonId = null;
-    });
-  }
-};
-export const addGeojsonSourceToLayer = (
-  centroids: DashboardGetProjectsData[] | undefined,
-  map: mapboxgl.Map,
-  layer: LayerType,
-  zoomFilterValue: number | undefined,
-  existsPolygons: boolean
-) => {
-  const { name, styles } = layer;
-  if (map && centroids && centroids.length > 0) {
-    if (map.getSource(name)) {
-      styles?.forEach((_: unknown, index: number) => {
-        map.removeLayer(`${name}-${index}`);
-      });
-      map.removeSource(name);
-    }
-
-    if (existsPolygons) {
-      return;
-    }
-
-    const features: GeoJSON.Feature[] = centroids.map((centroid: any) => ({
-      type: "Feature",
-      geometry: {
-        type: "Point",
-        coordinates: [centroid.long || centroid.centroid?.long, centroid.lat || centroid.centroid?.lat]
-      },
-      properties: {
-        uuid: centroid.uuid,
-        name: centroid.name || centroid.uuid,
-        type: centroid.type || "polygon_centroid"
-      }
-    }));
-
-    map.addSource(name, {
-      type: "geojson",
-      data: {
-        type: "FeatureCollection",
-        features: features
-      }
-    });
-
-    styles?.forEach((style: LayerWithStyle, index: number) => {
-      addLayerGeojsonStyle(map, name, name, style, index);
-    });
-  }
-};
-export const addSourceToLayer = (
-  layer: LayerType,
-  map: mapboxgl.Map,
-  polygonsData: Record<string, string[]> | undefined,
-  zoomFilter?: number | undefined,
-  isDashboard?: string | undefined
-) => {
-  const { name, geoserverLayerName, styles } = layer;
-  try {
-    if (map) {
-      if (map.getSource(name)) {
-        styles?.forEach((_: unknown, index: number) => {
-          map.removeLayer(`${name}-${index}`);
-        });
-        map.removeSource(name);
-      }
-      const GEOSERVER_TILE_URL = getGeoserverURL(geoserverLayerName, isDashboard);
-      map.addSource(name, {
-        type: "vector",
-        tiles: [GEOSERVER_TILE_URL]
-      });
-      styles?.forEach((style: LayerWithStyle, index: number) => {
-        addLayerStyle(map, name, geoserverLayerName, style, index, zoomFilter);
-      });
-      if (polygonsData) {
-        loadLayersInMap(map, polygonsData, layer, zoomFilter);
-      }
-    }
-  } catch (e) {
-    console.warn(e);
-  }
-};
-const loadDeleteLayer = (layer: any, map: mapboxgl.Map, polygonsData: Record<string, string[]> | undefined) => {
-  const { name, geoserverLayerName, styles } = layer;
-  styles?.forEach((style: LayerWithStyle, index: number) => {
-    if (map.getLayer(`${name}-${index}`)) {
-      map.removeLayer(`${name}-${index}`);
-    }
-    map.addLayer({
-      ...style,
-      id: `${name}-${index}`,
-      source: name,
-      "source-layer": geoserverLayerName
-    } as mapboxgl.AnyLayer);
-  });
-  loadLayersInMap(map, polygonsData, layer);
-};
-export const addDeleteLayer = (layer: any, map: mapboxgl.Map, polygonsData: Record<string, string[]> | undefined) => {
-  const { name, geoserverLayerName, styles } = layer;
-  try {
-    if (map) {
-      if (map.getSource(name)) {
-        styles?.forEach((_: unknown, index: number) => {
-          map.removeLayer(`${name}-${index}`);
-        });
-        map.removeSource(name);
-      }
-      const GEOSERVER_TILE_URL = getGeoserverURL(geoserverLayerName);
-      map.addSource(name, {
-        type: "vector",
-        tiles: [GEOSERVER_TILE_URL]
-      });
-      loadDeleteLayer(layer, map, polygonsData);
-    }
-  } catch (e) {
-    console.warn(e);
-  }
-};
-const moveDeleteLayers = (map: mapboxgl.Map) => {
-  const layers = layersList.filter(layer => layer.name === LAYERS_NAMES.DELETED_GEOMETRIES);
-  layers.forEach(layer => {
-    const { name, styles } = layer;
-    styles?.forEach((_: unknown, index: number) => {
-      if (map?.getLayer(`${name}-${index}`)) {
-        map?.moveLayer(`${name}-${index}`);
-      }
-    });
-  });
-};
-export const addLayerGeojsonStyle = (
-  map: mapboxgl.Map,
-  layerName: string,
-  sourceName: string,
-  style: LayerWithStyle,
-  index: number
-) => {
-  const beforeLayer = map.getLayer(LAYERS_NAMES.MEDIA_IMAGES) ? LAYERS_NAMES.MEDIA_IMAGES : undefined;
-  if (map.getLayer(`${layerName}-${index}`)) {
-    map.removeLayer(`${layerName}-${index}`);
-  }
-  map.addLayer(
-    {
-      ...style,
-      id: `${layerName}-${index}`,
-      source: sourceName
-    } as mapboxgl.AnyLayer,
-    beforeLayer
-  );
-  moveDeleteLayers(map);
-};
-export const setFilterCountry = (map: mapboxgl.Map, layerName: string, country: string) => {
-  const filter = ["==", ["get", "iso"], country];
-  map.setFilter(layerName, filter);
-};
-export const setFilterLandscape = (map: mapboxgl.Map, layerName: string, landscapes: string[]) => {
-  const filter = ["in", ["get", "landscape"], ["literal", landscapes]];
-  map.setFilter(layerName, filter);
-};
-export const addBorderCountry = (map: mapboxgl.Map, country: string) => {
-  if (!country || !map) return;
-  try {
-    const styleName = `${LAYERS_NAMES.WORLD_COUNTRIES}-line`;
-    const countryLayer = layersList.find(layer => layer.name === styleName);
-    if (!countryLayer) return;
-    const countryStyles = countryLayer.styles || [];
-    const sourceName = countryLayer.name;
-    const GEOSERVER_TILE_URL = getGeoserverURL(countryLayer.geoserverLayerName);
-
-    if (!map.getSource(sourceName)) {
-      map.addSource(sourceName, {
-        type: "vector",
-        tiles: [GEOSERVER_TILE_URL]
-      });
-    }
-    if (map.getLayer(sourceName)) {
-      map.removeLayer(sourceName);
-    }
-    const style = countryStyles[0];
-    map.addLayer({
-      ...style,
-      id: sourceName,
-      source: sourceName,
-      "source-layer": countryLayer.geoserverLayerName
-    } as mapboxgl.AnyLayer);
-    setFilterCountry(map, sourceName, country);
-  } catch (e) {
-    console.warn(e);
-  }
-};
-export const addBorderLandscape = (map: mapboxgl.Map, landscapes: string[]) => {
-  if (!landscapes || !map) return;
-
-  const styleName = `${LAYERS_NAMES.LANDSCAPES}`;
-  const landscapeLayer = layersList.find(layer => layer.name === styleName);
-  if (!landscapeLayer) return;
-  const countryStyles = landscapeLayer.styles || [];
-  const sourceName = landscapeLayer.name;
-  const GEOSERVER_TILE_URL = getGeoserverURL(landscapeLayer.geoserverLayerName);
-
-  if (!map.getSource(sourceName)) {
-    map.addSource(sourceName, {
-      type: "vector",
-      tiles: [GEOSERVER_TILE_URL]
-    });
-  }
-  if (map.getLayer(sourceName)) {
-    map.removeLayer(sourceName);
-  }
-  const style = countryStyles[0];
-  map.addLayer({
-    ...style,
-    id: sourceName,
-    source: sourceName,
-    "source-layer": landscapeLayer.geoserverLayerName
-  } as mapboxgl.AnyLayer);
-  setFilterLandscape(map, sourceName, landscapes);
-};
-export const removeBorderLandscape = (map: mapboxgl.Map) => {
-  if (!map || !map.isStyleLoaded()) return;
-
-  const layerName = `${LAYERS_NAMES.LANDSCAPES}`;
-  try {
-    if (map.getLayer(layerName)) {
-      map.removeLayer(layerName);
-    }
-    if (map.getSource(layerName)) {
-      map.removeSource(layerName);
-    }
-  } catch (error) {
-    Log.warn("Error removing border landscape:", error);
-  }
-};
-
-export const removeBorderCountry = (map: mapboxgl.Map) => {
-  if (!map || !map.isStyleLoaded()) return;
-
-  const layerName = `${LAYERS_NAMES.WORLD_COUNTRIES}-line`;
-  try {
-    if (map.getLayer(layerName)) {
-      map.removeLayer(layerName);
-    }
-    if (map.getSource(layerName)) {
-      map.removeSource(layerName);
-    }
-  } catch (error) {
-    Log.warn("Error removing border country:", error);
-  }
-};
-
-export const addLayerStyle = (
-  map: mapboxgl.Map,
-  layerName: string,
-  sourceName: string,
-  style: LayerWithStyle,
-  index_suffix: number | string,
-  zoomFilter?: number | undefined
-) => {
-  const beforeLayer = map.getLayer(LAYERS_NAMES.MEDIA_IMAGES) ? LAYERS_NAMES.MEDIA_IMAGES : undefined;
-  if (map.getLayer(`${layerName}-${index_suffix}`)) {
-    map.removeLayer(`${layerName}-${index_suffix}`);
-  }
-  map.addLayer(
-    {
-      ...style,
-      id: `${layerName}-${index_suffix}`,
-      source: sourceName,
-      "source-layer": sourceName,
-      ...(zoomFilter && {
-        filter: ["all", style.filter || ["==", true, true], [">=", ["zoom"], zoomFilter]]
-      })
-    } as mapboxgl.AnyLayer,
-    beforeLayer
-  );
-  moveDeleteLayers(map);
-};
-
-export const updateMapProjection = (map: mapboxgl.Map, currentStyle: MapStyle) => {
-  const config = BASEMAP_CONFIGS[currentStyle];
-  if (config?.projection) {
-    map.setProjection(config.projection);
-  }
-};
-
-const GOOGLE_RASTER_SOURCE_ID = "google-satellite-source";
-const GOOGLE_RASTER_LAYER_ID = "google-satellite-layer";
-
-export const addGoogleSatelliteLayer = (map: mapboxgl.Map) => {
-  if (!map || !map.isStyleLoaded()) {
-    Log.warn("Map or style not loaded, cannot add Google satellite layer");
-    return;
-  }
-
-  removeGoogleSatelliteLayer(map);
-
-  const config = BASEMAP_CONFIGS[MapStyle.GoogleSatellite];
-  if (!config.rasterUrl) {
-    Log.warn("Google satellite raster URL not configured");
-    return;
-  }
-
-  try {
-    const layers = map.getStyle().layers ?? [];
-    const polygonLayerPrefixes = [
-      LAYERS_NAMES.POLYGON_GEOMETRY,
-      LAYERS_NAMES.DELETED_GEOMETRIES,
-      LAYERS_NAMES.CENTROIDS,
-      LAYERS_NAMES.POLYGON_CENTROIDS,
-      ANR_PLOT_LAYER_PREFIX
-    ];
-    const isPolygonLayer = (layerId: string) => {
-      return polygonLayerPrefixes.some(prefix => layerId.startsWith(prefix));
-    };
-
-    const layersToHide = layers.filter(
-      layer =>
-        (layer.type === "background" ||
-          (layer.type === "fill" && !isPolygonLayer(layer.id)) ||
-          (layer.type === "raster" && layer.id !== GOOGLE_RASTER_LAYER_ID)) &&
-        !isPolygonLayer(layer.id)
-    );
-
-    let hiddenCount = 0;
-    layersToHide.forEach(layer => {
-      try {
-        const currentVisibility = map.getLayoutProperty(layer.id, "visibility");
-        if (currentVisibility !== "none") {
-          map.setLayoutProperty(layer.id, "visibility", "none");
-          hiddenCount++;
-        }
-      } catch (e) {
-        Log.warn("Error setting layer visibility:", e);
-      }
-    });
-
-    if (hiddenCount > 0) {
-      Log.info(`Hidden ${hiddenCount} Street layers`);
-    }
-
-    map.addSource(GOOGLE_RASTER_SOURCE_ID, {
-      type: "raster",
-      tiles: [config.rasterUrl],
-      tileSize: 256,
-      minzoom: 0,
-      maxzoom: 22
-    });
-
-    let firstLayerId: string | undefined;
-    for (const layer of layers) {
-      if (layer.type !== "background") {
-        firstLayerId = layer.id;
-        break;
-      }
-    }
-
-    map.addLayer(
-      {
-        id: GOOGLE_RASTER_LAYER_ID,
-        type: "raster",
-        source: GOOGLE_RASTER_SOURCE_ID,
-        paint: {
-          "raster-opacity": 1
-        }
-      },
-      firstLayerId
-    );
-  } catch (error) {
-    Log.error("Error adding Google satellite layer:", error);
-  }
-};
-
-export const removeGoogleSatelliteLayer = (map: mapboxgl.Map) => {
-  if (!map) return;
-  if (!map.isStyleLoaded()) {
-    return;
-  }
-
-  try {
-    if (map.getLayer(GOOGLE_RASTER_LAYER_ID)) {
-      map.removeLayer(GOOGLE_RASTER_LAYER_ID);
-    }
-    if (map.getSource(GOOGLE_RASTER_SOURCE_ID)) {
-      map.removeSource(GOOGLE_RASTER_SOURCE_ID);
-    }
-
-    const layers = map.getStyle()?.layers ?? [];
-    const polygonLayerPrefixes = [
-      LAYERS_NAMES.POLYGON_GEOMETRY,
-      LAYERS_NAMES.DELETED_GEOMETRIES,
-      LAYERS_NAMES.CENTROIDS,
-      LAYERS_NAMES.POLYGON_CENTROIDS,
-      ANR_PLOT_LAYER_PREFIX
-    ];
-    const isPolygonLayer = (layerId: string) => {
-      return polygonLayerPrefixes.some(prefix => layerId.startsWith(prefix));
-    };
-
-    const layersToRestore = layers.filter(layer => {
-      const isBaseMapLayer = layer.type === "background" || layer.type === "fill" || layer.type === "raster";
-      return isBaseMapLayer && !isPolygonLayer(layer.id);
-    });
-
-    let restoredCount = 0;
-    layersToRestore.forEach(layer => {
-      try {
-        const currentVisibility = map.getLayoutProperty(layer.id, "visibility");
-        if (currentVisibility === "none") {
-          map.setLayoutProperty(layer.id, "visibility", "visible");
-          restoredCount++;
-        }
-      } catch (e) {
-        Log.warn("Error removing Google satellite layer:", e);
-      }
-    });
-
-    if (restoredCount > 0) {
-      Log.info(`Restored ${restoredCount} Street layers`);
-    }
-  } catch (error) {
-    Log.warn("Error removing Google satellite layer:", error);
-  }
-};
-
-const isValidGeographicBBox = (bbox: BBox): boolean => {
-  const [minLng, minLat, maxLng, maxLat] = bbox;
-
-  const isValidLng = minLng >= -180 && minLng <= 180 && maxLng >= -180 && maxLng <= 180;
-  const isValidLat = minLat >= -90 && minLat <= 90 && maxLat >= -90 && maxLat <= 90;
-  const isOrdered = minLng < maxLng && minLat < maxLat;
-  return isValidLng && isValidLat && isOrdered;
-};
-
-export const zoomToBbox = (bbox: BBox, map: mapboxgl.Map, hasControls: boolean, currentStyle = MapStyle.Satellite) => {
-  if (!map || !bbox) {
-    return;
-  }
-
-  if (!isValidGeographicBBox(bbox)) {
-    Log.warn(
-      "zoomToBbox: Invalid geographic coordinates detected. Expected longitude between -180/180 and latitude between -90/90, but received:",
-      bbox
-    );
-    return;
-  }
-
-  try {
-    map.fitBounds(bbox, {
-      padding: hasControls ? 100 : 30,
-      linear: false,
-      animate: true
-    });
-  } catch (error) {
-    Log.warn("zoomToBbox: Error occurred while fitting bounds:", error);
-    return;
-  }
-};
-
-export const zoomToCenter = (center: [number, number], zoom: number, map: mapboxgl.Map) => {
-  if (!map || !center || zoom === undefined) {
-    return;
-  }
-
-  const [lng, lat] = center;
-
-  if (lng < -180 || lng > 180 || lat < -90 || lat > 90) {
-    Log.warn(
-      "zoomToCenter: Invalid geographic coordinates detected. Expected longitude between -180/180 and latitude between -90/90, but received:",
-      center
-    );
-    return;
-  }
-
-  if (zoom < 0 || zoom > 22) {
-    Log.warn("zoomToCenter: Invalid zoom level. Expected between 0 and 22, but received:", zoom);
-    return;
-  }
-
-  try {
-    map.setCenter([lng, lat]);
-    map.setZoom(zoom);
-  } catch (error) {
-    Log.warn("zoomToCenter: Error occurred while setting center and zoom:", error);
-    return;
-  }
-};
 
 export const formatPlannedStartDate = (plantStartDate: Date | null | undefined): string => {
   return plantStartDate != null
@@ -1217,48 +60,7 @@ export const formatCommentaryDate = (date: Date | null | undefined): string => {
     : "Unknown";
 };
 
-export function parsePolygonDataV3(sitePolygonData: SitePolygonLightDto[] | undefined) {
-  return (sitePolygonData ?? []).reduce((acc: Record<string, string[]>, data: SitePolygonLightDto) => {
-    if (data.status != null && data.polygonUuid != null) {
-      if (acc[data.status] == null) {
-        acc[data.status] = [];
-      }
-      acc[data.status].push(data.polygonUuid);
-    }
-    return acc;
-  }, {});
-}
-
-export const countStatusesV3 = (sitePolygonData: SitePolygonLightDto[]): DataPolygonOverview => {
-  const statusOrder = ["Draft", "Submitted", "Needs Info", "Approved"];
-
-  const statusCountMap: Record<string, number> = {};
-
-  sitePolygonData.forEach(item => {
-    let statusKey = item.status?.toLowerCase();
-
-    if (statusKey) {
-      if (statusKey === "needs-more-information") {
-        statusKey = "Needs Info";
-      } else {
-        statusKey = statusKey.replace(/\b\w/g, char => char.toUpperCase());
-      }
-
-      statusCountMap[statusKey] = (statusCountMap[statusKey] || 0) + 1;
-    }
-  });
-
-  const unorderedData = Object.entries(statusCountMap).map(([status, count]) => ({
-    status,
-    count
-  }));
-
-  const orderedData = unorderedData.sort((a, b) => statusOrder.indexOf(a.status) - statusOrder.indexOf(b.status));
-
-  return orderedData;
-};
-
-export const formatFileName = (inputString: string) => {
+export const formatFileName = (inputString: string): string => {
   return inputString.toLowerCase().replace(/\s+/g, "_");
 };
 
@@ -1266,18 +68,13 @@ export const extractGeoJsonFromResponse = (
   response: GeoJsonExportDto | { data?: { attributes?: GeoJsonExportDto } } | undefined
 ): GeoJSON.FeatureCollection | undefined => {
   if (!response) return undefined;
-
-  if ("type" in response && response.type === "FeatureCollection") {
+  if ("type" in response && (response as any).type === "FeatureCollection") {
     return response as unknown as GeoJSON.FeatureCollection;
   }
-
-  if ("data" in response && response.data?.attributes) {
-    const attributes = response.data.attributes;
-    if (attributes.type === "FeatureCollection") {
-      return attributes as unknown as GeoJSON.FeatureCollection;
-    }
+  if ("data" in response && (response as any).data?.attributes) {
+    const attributes = (response as any).data.attributes;
+    if (attributes.type === "FeatureCollection") return attributes as unknown as GeoJSON.FeatureCollection;
   }
-
   return undefined;
 };
 
@@ -1306,18 +103,10 @@ export async function downloadPolygonGeoJson(
   options?: Omit<GetSitePolygonsGeoJsonQueryParams, "uuid" | "siteUuid">
 ): Promise<void> {
   try {
-    const result = await loadPolygonGeoJson({
-      uuid: polygonUuid,
-      ...options
-    });
-
+    const result = await loadPolygonGeoJson({ uuid: polygonUuid, ...options });
     const geojson = extractGeoJsonFromResponse(result.data);
-    if (!geojson) {
-      throw new Error("Failed to extract GeoJSON from response");
-    }
-
-    const safeFilename = formatFileName(filename ?? polygonUuid);
-    downloadGeoJsonFile(geojson, safeFilename);
+    if (!geojson) throw new Error("Failed to extract GeoJSON from response");
+    downloadGeoJsonFile(geojson, formatFileName(filename ?? polygonUuid));
   } catch (error) {
     Log.error("Failed to download polygon GeoJSON:", error);
     throw error;
@@ -1330,18 +119,10 @@ export async function downloadProjectSitePolygonsGeoJson(
   options?: Omit<GetSitePolygonsGeoJsonQueryParams, "uuid" | "projectUuid">
 ): Promise<void> {
   try {
-    const result = await loadProjectSitePolygonsGeoJson({
-      projectUuid,
-      ...options
-    });
-
+    const result = await loadProjectSitePolygonsGeoJson({ projectUuid, ...options });
     const geojson = extractGeoJsonFromResponse(result.data);
-    if (!geojson) {
-      throw new Error("Failed to extract GeoJSON from response");
-    }
-
-    const safeFilename = formatFileName(projectName);
-    downloadGeoJsonFile(geojson, safeFilename);
+    if (!geojson) throw new Error("Failed to extract GeoJSON from response");
+    downloadGeoJsonFile(geojson, formatFileName(projectName));
   } catch (error) {
     Log.error("Failed to download project site polygons GeoJSON:", error);
     throw error;
@@ -1354,60 +135,12 @@ export async function downloadSitePolygonsGeoJson(
   options?: Omit<GetSitePolygonsGeoJsonQueryParams, "uuid" | "siteUuid">
 ): Promise<void> {
   try {
-    const result = await loadSitePolygonsGeoJson({
-      siteUuid,
-      ...options
-    });
-
+    const result = await loadSitePolygonsGeoJson({ siteUuid, ...options });
     const geojson = extractGeoJsonFromResponse(result.data);
-    if (!geojson) {
-      throw new Error("Failed to extract GeoJSON from response");
-    }
-
-    const safeFilename = formatFileName(siteName);
-    downloadGeoJsonFile(geojson, safeFilename);
+    if (!geojson) throw new Error("Failed to extract GeoJSON from response");
+    downloadGeoJsonFile(geojson, formatFileName(siteName));
   } catch (error) {
     Log.error("Failed to download site polygons GeoJSON:", error);
-    throw error;
-  }
-}
-
-export async function fetchPolygonGeometry(
-  polygonUuid: string,
-  geometryOnly: boolean = true,
-  projectPitchUuid?: string
-): Promise<any | null> {
-  if (polygonUuid == null || polygonUuid === "") {
-    Log.error("fetchPolygonGeometry called with undefined or empty polygonUuid");
-    throw new Error("polygonUuid is required");
-  }
-
-  try {
-    let result;
-
-    if (projectPitchUuid != null) {
-      result = await loadProjectPolygonsGeoJson({
-        projectPitchUuid,
-        enabled: true
-      });
-    } else {
-      result = await loadPolygonGeoJson({
-        uuid: polygonUuid,
-        geometryOnly,
-        includeExtendedData: false,
-        enabled: true
-      });
-    }
-
-    const geojson = extractGeoJsonFromResponse(result.data);
-    if (geojson == null || geojson.features == null || geojson.features.length === 0) {
-      Log.warn("No geometry found in GeoJSON response for polygon:", polygonUuid);
-      return null;
-    }
-
-    return geojson.features[0].geometry;
-  } catch (error) {
-    Log.error("Failed to fetch polygon geometry:", error);
     throw error;
   }
 }
@@ -1417,35 +150,19 @@ export async function fetchMultiplePolygonsGeoJson(
   includeExtendedData: boolean = true
 ): Promise<GeoJSON.FeatureCollection> {
   try {
-    const promises = polygonUuids.map(uuid =>
-      loadPolygonGeoJson({
-        uuid,
-        includeExtendedData
-      })
-    );
-
-    const results = await Promise.all(promises);
+    const results = await Promise.all(polygonUuids.map(uuid => loadPolygonGeoJson({ uuid, includeExtendedData })));
     const features: GeoJSON.Feature[] = [];
-
     results.forEach((result, index) => {
       const geojson = extractGeoJsonFromResponse(result.data);
-      if (geojson && geojson.features) {
+      if (geojson?.features) {
         geojson.features.forEach(feature => {
-          if (!feature.properties) {
-            feature.properties = {};
-          }
-          if (!feature.properties.uuid) {
-            feature.properties.uuid = polygonUuids[index];
-          }
+          if (!feature.properties) feature.properties = {};
+          if (!feature.properties.uuid) feature.properties.uuid = polygonUuids[index];
           features.push(feature);
         });
       }
     });
-
-    return {
-      type: "FeatureCollection",
-      features
-    };
+    return { type: "FeatureCollection", features };
   } catch (error) {
     Log.error("Failed to fetch multiple polygons GeoJSON:", error);
     throw error;
@@ -1459,13 +176,10 @@ export async function downloadMultiplePolygonsGeoJson(
 ): Promise<void> {
   try {
     const combinedGeojson = await fetchMultiplePolygonsGeoJson(polygonUuids, includeExtendedData);
-
     if (!combinedGeojson.features || combinedGeojson.features.length === 0) {
       throw new Error("No polygons found to download");
     }
-
-    const safeFilename = formatFileName(filename);
-    downloadGeoJsonFile(combinedGeojson, safeFilename);
+    downloadGeoJsonFile(combinedGeojson, formatFileName(filename));
   } catch (error) {
     Log.error("Failed to download multiple polygons GeoJSON:", error);
     throw error;
@@ -1482,268 +196,16 @@ export async function downloadProjectPolygonsGeoJson(
   options?: Omit<GetSitePolygonsGeoJsonQueryParams, "uuid" | "siteUuid" | "projectUuid">
 ): Promise<void> {
   try {
-    const result = await loadProjectPolygonsGeoJson({
-      projectPitchUuid,
-      ...options
-    });
-
+    const result = await loadProjectPolygonsGeoJson({ projectPitchUuid, ...options });
     const geojson = extractGeoJsonFromResponse(result.data);
-    if (!geojson) {
-      throw new Error("Failed to extract GeoJSON from response");
-    }
-
-    if (!geojson.features || geojson.features.length === 0) {
-      throw new Error("No polygons found to download");
-    }
-
-    const safeFilename = formatFileName(projectName);
-    downloadGeoJsonFile(geojson, `${safeFilename}_polygons`);
+    if (!geojson) throw new Error("Failed to extract GeoJSON from response");
+    if (!geojson.features || geojson.features.length === 0) throw new Error("No polygons found to download");
+    downloadGeoJsonFile(geojson, `${formatFileName(projectName)}_polygons`);
   } catch (error) {
     Log.error("Failed to download project polygons GeoJSON:", error);
     throw error;
   }
 }
-
-export async function storePolygon(
-  geojson: any,
-  record: any,
-  setPolygonFromMap?: any,
-  refetchSitePolygons?: () => any
-) {
-  if (geojson?.length) {
-    const attributes: CreateSitePolygonAttributesDto = {
-      geometries: [
-        {
-          type: "FeatureCollection",
-          features: [
-            {
-              type: "Feature",
-              geometry: geojson[0].geometry,
-              properties: {
-                site_id: record.uuid
-              }
-            }
-          ] as any
-        }
-      ]
-    };
-
-    try {
-      const result = await createSitePolygonsResource(attributes);
-
-      if (refetchSitePolygons) {
-        await refetchSitePolygons();
-      }
-      if (setPolygonFromMap) {
-        setPolygonFromMap({ uuid: result.polygonUuid, isOpen: true, primary_uuid: result.primaryUuid });
-      }
-    } catch (error) {
-      console.error("Failed to create site polygon:", error);
-      throw error;
-    }
-  }
-}
-
-export async function storePolygonProject(
-  geojson: any,
-  entityUuid: string,
-  entityType: string,
-  refetch: any,
-  setPolygonFromMap: any
-) {
-  if (geojson?.length) {
-    const geometries = [
-      {
-        type: "FeatureCollection",
-        features: geojson.map((feature: any) => ({
-          type: "Feature",
-          geometry: feature.geometry,
-          properties: {
-            projectPitchUuid: entityUuid
-          }
-        }))
-      }
-    ];
-
-    const response = await createProjectPolygonWithReplace({ geometries }, entityUuid);
-
-    const polygonUuid = response.polygonUuid;
-    if (polygonUuid) {
-      refetch?.();
-      setPolygonFromMap?.({
-        uuid: polygonUuid,
-        isOpen: true,
-        entityName: "project-pitches",
-        projectPitchUuid: entityUuid
-      });
-    }
-  }
-}
-
-export async function updatePolygonProjectGeometry(geojson: any, polygonUuid: string, refetch: any) {
-  if (geojson?.length && polygonUuid) {
-    const geometries = [
-      {
-        type: "FeatureCollection",
-        features: geojson.map((feature: any) => ({
-          type: "Feature",
-          geometry: feature.geometry,
-          properties: {}
-        }))
-      }
-    ];
-
-    await updateProjectPolygonResource(polygonUuid, { geometries });
-
-    const refetchResult = refetch?.();
-    if (refetchResult && typeof refetchResult.then === "function") {
-      await refetchResult;
-    }
-  }
-}
-export const drawTemporaryPolygon = (geojson: any, cb: Function, map: mapboxgl.Map, polygonVersion?: any) => {
-  if (geojson) {
-    const geojsonFormatted = convertToAcceptedGEOJSON(geojson);
-    if (polygonVersion?.poly_id && !polygonVersion?.is_active) {
-      map.addSource("temp-polygon-source", {
-        type: "geojson",
-        data: geojsonFormatted
-      });
-
-      map?.addLayer({
-        id: `temp-polygon-source`,
-        type: "fill",
-        source: `temp-polygon-source`,
-        layout: {},
-        paint: {
-          "fill-color": getPolygonColor(polygonVersion?.status),
-          "fill-opacity": 0
-        }
-      });
-      map.addLayer({
-        id: `temp-polygon-source-line`,
-        type: "line",
-        source: `temp-polygon-source`,
-        layout: {},
-        paint: {
-          "line-color": getPolygonColor(polygonVersion?.status),
-          "line-width": 2,
-          "line-dasharray": [4, 2]
-        }
-      });
-    }
-    zoomToBbox(bbox(geojsonFormatted) as BBox, map, false);
-    cb(polygonVersion?.poly_id as string);
-  }
-};
-
-const getPolygonColor = (polygonStatus: string) => {
-  switch (polygonStatus) {
-    case "draft":
-      return "#E468EF";
-    case "submitted":
-      return "#2398d8";
-    case "approved":
-      return "#72d961";
-    case "needs-more-information":
-      return "#ff8938";
-    default:
-      return "#000000";
-  }
-};
-
-export const addMarkerAndZoom = (map: mapboxgl.Map, location: { lng: number; lat: number }) => {
-  if (map) {
-    const { lng, lat } = location;
-    const lngLat = new mapboxgl.LngLat(lng, lat);
-    createMarker(lngLat, map);
-    map.setCenter([lng, lat]);
-    map.setZoom(14);
-  }
-};
-
-export const createMarker = (lngLat: LngLat, map: mapboxgl.Map) => {
-  return new mapboxgl.Marker({
-    color: "#ba5856"
-  })
-    .setLngLat(lngLat)
-    .addTo(map);
-};
-
-export const getCurrentMapStyle = (map: mapboxgl.Map): MapStyle | undefined => {
-  if (!map) return undefined;
-  try {
-    if (map.getLayer(GOOGLE_RASTER_LAYER_ID)) {
-      return MapStyle.GoogleSatellite;
-    }
-
-    const styleUrl = (map as any).style?.url || (map as any).style?.name;
-    if (styleUrl) {
-      if (styleUrl === MapStyle.Street || styleUrl.includes("clve3yxzq01w101pefkkg3rci")) {
-        return MapStyle.Street;
-      }
-      if (styleUrl === MapStyle.Satellite || styleUrl.includes("clv3bkxut01y301pk317z5afu")) {
-        return MapStyle.Satellite;
-      }
-    }
-  } catch (e) {
-    Log.warn("Error getting current map style:", e);
-  }
-  return undefined;
-};
-
-export const setMapStyle = (
-  targetStyle: MapStyle,
-  map: mapboxgl.Map,
-  setCurrentStyle: (style: MapStyle) => void,
-  currentStyle: string | MapStyle
-) => {
-  if (!map || currentStyle === targetStyle) return;
-
-  const targetConfig = BASEMAP_CONFIGS[targetStyle];
-
-  if (targetStyle === MapStyle.GoogleSatellite) {
-    setCurrentStyle(targetStyle);
-
-    if (currentStyle === MapStyle.Street) {
-      if (map.isStyleLoaded()) {
-        addGoogleSatelliteLayer(map);
-        updateMapProjection(map, targetStyle);
-      } else {
-        map.once("style.load", () => {
-          addGoogleSatelliteLayer(map);
-          updateMapProjection(map, targetStyle);
-        });
-      }
-    } else {
-      map.setStyle(targetConfig.style);
-    }
-    return;
-  }
-
-  if (currentStyle === MapStyle.GoogleSatellite) {
-    if (targetStyle === MapStyle.Street) {
-      if (map.isStyleLoaded()) {
-        removeGoogleSatelliteLayer(map);
-        setCurrentStyle(targetStyle);
-        updateMapProjection(map, targetStyle);
-      } else {
-        map.once("style.load", () => {
-          removeGoogleSatelliteLayer(map);
-          setCurrentStyle(targetStyle);
-          updateMapProjection(map, targetStyle);
-        });
-      }
-      return;
-    }
-  }
-
-  setCurrentStyle(targetStyle);
-  map.setStyle(targetConfig.style);
-  map.once("style.load", () => {
-    updateMapProjection(map, targetStyle);
-  });
-};
 
 export type ValidationRecordV3 = {
   uuid: string;
@@ -1759,9 +221,7 @@ export function parseValidationDataV3(
 ) {
   const validationMap = new Map<string, ValidationRecordV3>();
   currentValidationSite.forEach(validation => {
-    if (validation?.uuid != null) {
-      validationMap.set(validation.uuid, validation);
-    }
+    if (validation?.uuid != null) validationMap.set(validation.uuid, validation);
   });
 
   return (sitePolygonData ?? []).map(site => {
@@ -1778,198 +238,4 @@ export function parseValidationDataV3(
       ...(polygonValidation.length > 0 ? { polygonValidation } : {})
     };
   });
-}
-
-type AnrPlotOverlayState = {
-  clickHandler: ((e: mapboxgl.MapLayerMouseEvent) => void) | null;
-  mouseEnterHandler: (() => void) | null;
-  mouseLeaveHandler: (() => void) | null;
-  popup: mapboxgl.Popup | null;
-  pendingIdleRetry: { fn: () => void } | null;
-};
-
-const anrPlotOverlayStateByMap = new WeakMap<mapboxgl.Map, AnrPlotOverlayState>();
-
-function getAnrPlotOverlayState(map: mapboxgl.Map): AnrPlotOverlayState {
-  const existing = anrPlotOverlayStateByMap.get(map);
-  if (existing != null) return existing;
-  const created: AnrPlotOverlayState = {
-    clickHandler: null,
-    mouseEnterHandler: null,
-    mouseLeaveHandler: null,
-    popup: null,
-    pendingIdleRetry: null
-  };
-  anrPlotOverlayStateByMap.set(map, created);
-  return created;
-}
-
-function cancelAnrPendingRetry(map: mapboxgl.Map) {
-  const state = getAnrPlotOverlayState(map);
-  if (state.pendingIdleRetry != null) {
-    map.off("idle", state.pendingIdleRetry.fn);
-    state.pendingIdleRetry = null;
-  }
-}
-
-export function removeAnrPlotGeometryOverlay(map: mapboxgl.Map | null | undefined) {
-  if (map == null) return;
-  cancelAnrPendingRetry(map);
-  const state = getAnrPlotOverlayState(map);
-  try {
-    if (state.popup != null) {
-      state.popup.remove();
-      state.popup = null;
-    }
-    if (state.clickHandler != null) {
-      map.off("click", ANR_PLOT_FILL_LAYER_ID, state.clickHandler);
-      state.clickHandler = null;
-    }
-    if (state.mouseEnterHandler != null) {
-      map.off("mouseenter", ANR_PLOT_FILL_LAYER_ID, state.mouseEnterHandler);
-      state.mouseEnterHandler = null;
-    }
-    if (state.mouseLeaveHandler != null) {
-      map.off("mouseleave", ANR_PLOT_FILL_LAYER_ID, state.mouseLeaveHandler);
-      state.mouseLeaveHandler = null;
-    }
-    if (map.getLayer(ANR_PLOT_LINE_LAYER_ID) != null) {
-      map.removeLayer(ANR_PLOT_LINE_LAYER_ID);
-    }
-    if (map.getLayer(ANR_PLOT_FILL_LAYER_ID) != null) {
-      map.removeLayer(ANR_PLOT_FILL_LAYER_ID);
-    }
-    if (map.getSource(ANR_PLOT_SOURCE_ID) != null) {
-      map.removeSource(ANR_PLOT_SOURCE_ID);
-    }
-  } catch (e) {
-    Log.warn("removeAnrPlotGeometryOverlay:", e);
-  }
-}
-
-export function upsertAnrPlotGeometryOverlay(map: mapboxgl.Map, geojson: unknown, options: { visible: boolean }) {
-  if (map == null) return;
-  removeAnrPlotGeometryOverlay(map);
-  const state = getAnrPlotOverlayState(map);
-
-  if (!options.visible) {
-    return;
-  }
-
-  const geojsonFormatted = convertToAcceptedGEOJSON(geojson) as GeoJSON.FeatureCollection;
-  if (geojsonFormatted?.features == null || geojsonFormatted.features.length === 0) {
-    return;
-  }
-
-  if (!map.isStyleLoaded()) {
-    const retryFn = () => {
-      const currentState = getAnrPlotOverlayState(map);
-      currentState.pendingIdleRetry = null;
-      upsertAnrPlotGeometryOverlay(map, geojson, options);
-    };
-    state.pendingIdleRetry = { fn: retryFn };
-    map.once("idle", retryFn);
-    return;
-  }
-
-  const beforeLayer = map.getLayer(LAYERS_NAMES.MEDIA_IMAGES) != null ? LAYERS_NAMES.MEDIA_IMAGES : undefined;
-
-  try {
-    map.addSource(ANR_PLOT_SOURCE_ID, {
-      type: "geojson",
-      data: geojsonFormatted as GeoJSON.FeatureCollection
-    });
-
-    map.addLayer(
-      {
-        id: ANR_PLOT_FILL_LAYER_ID,
-        type: "fill",
-        source: ANR_PLOT_SOURCE_ID,
-        layout: { visibility: "visible" },
-        paint: {
-          "fill-color": "#9ca3af",
-          "fill-opacity": 0.7
-        }
-      },
-      beforeLayer
-    );
-
-    map.addLayer(
-      {
-        id: ANR_PLOT_LINE_LAYER_ID,
-        type: "line",
-        source: ANR_PLOT_SOURCE_ID,
-        layout: { visibility: "visible" },
-        paint: {
-          "line-color": "#6b7280",
-          "line-width": 1.95,
-          "line-opacity": 0.9
-        }
-      },
-      beforeLayer
-    );
-
-    if (map.getLayer(LAYERS_NAMES.MEDIA_IMAGES) != null) {
-      try {
-        map.moveLayer(ANR_PLOT_LINE_LAYER_ID, LAYERS_NAMES.MEDIA_IMAGES);
-        map.moveLayer(ANR_PLOT_FILL_LAYER_ID, ANR_PLOT_LINE_LAYER_ID);
-      } catch (e) {
-        Log.warn("moveLayer ANR plot overlay:", e);
-      }
-    }
-    state.clickHandler = (e: mapboxgl.MapLayerMouseEvent) => {
-      const feature = e.features?.[0];
-      if (feature == null) return;
-      if (state.popup != null) {
-        state.popup.remove();
-        state.popup = null;
-      }
-      const props = feature.properties ?? {};
-      const popupContent = document.createElement("div");
-      popupContent.className = "popup-content-map";
-      const root = createRoot(popupContent);
-      const rawPlotId = props.plotId;
-      const rawArea = props.areaM2;
-      const plotIdParsed =
-        typeof rawPlotId === "number"
-          ? rawPlotId
-          : rawPlotId != null && !Number.isNaN(Number(rawPlotId))
-          ? Number(rawPlotId)
-          : undefined;
-      const areaParsed =
-        typeof rawArea === "number"
-          ? rawArea
-          : rawArea != null && !Number.isNaN(Number(rawArea))
-          ? Number(rawArea)
-          : undefined;
-      root.render(
-        createElement(AnrPlotMapPopup, {
-          plotId: plotIdParsed,
-          areaM2: areaParsed,
-          select: props.select != null ? String(props.select) : null,
-          onClose: () => {
-            state.popup?.remove();
-            state.popup = null;
-          }
-        })
-      );
-      state.popup = new mapboxgl.Popup({ className: "popup-map", closeButton: false })
-        .setLngLat(e.lngLat)
-        .setDOMContent(popupContent)
-        .addTo(map);
-    };
-
-    map.on("click", ANR_PLOT_FILL_LAYER_ID, state.clickHandler);
-
-    state.mouseEnterHandler = () => {
-      map.getCanvas().style.cursor = "pointer";
-    };
-    state.mouseLeaveHandler = () => {
-      map.getCanvas().style.cursor = "";
-    };
-    map.on("mouseenter", ANR_PLOT_FILL_LAYER_ID, state.mouseEnterHandler);
-    map.on("mouseleave", ANR_PLOT_FILL_LAYER_ID, state.mouseLeaveHandler);
-  } catch (e) {
-    Log.warn("upsertAnrPlotGeometryOverlay:", e);
-  }
 }
